Format diffguard utility and simplify docs
